+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-- No changes yet
+- Do not require constant link waiting time attribute to be set for dynamic waiting time estimation
+- Add 'null' subpopulation scoring parameters by default
+- Automatically pass vehicle type to generators
 
 **1.0.1**